requirements.txt
numpy
pillow
loremipsum
names
coolname
pyhive